improve perf
